requirements.txt
customtkinter
mss
numpy
opencv-python
pydirectinput-rgx
Pillow
pygetwindow
requests
pynput
